Adjust hidden analysis visit ratio in AnalysisView
Change the calculation of the `isHidden` variable in the `AnalysisView` struct to utilize the `hiddenAnalysisVisitRatio` value from the `ConfigView` struct.

Previously, the `isHidden` value was calculated using a fixed threshold of 0.1 multiplied by the maximum number of visits. Now, it is based on `config.hiddenAnalysisVisitRatio` multiplied by the maximum number of visits.

This change allows for greater flexibility in controlling the visibility of analysis views based on the desired hidden visit ratio. It provides a more customizable and dynamic user experience.