Improve performance of changing board sizes
This change updates the `ConfigView` in order to send the appropriate KataGo command for changing the board size when the edit mode is inactive and the board size is modified. By checking the `isBoardSizeChanged` property, we ensure that the command is only sent once per change. This improvement enhances the performance by accurately reflecting the changed board size in the game analysis.